Add Password change , password reset , email update feature

+++ app/src/main/java/com/aditya/socialguru/ui_layer/fragment/profile_part/SettingFragment.kt
@@ -9,6 +9,9 @@ import androidx.coordinatorlayout.widget.CoordinatorLayout
 import androidx.fragment.app.Fragment
 import androidx.fragment.app.viewModels
 import androidx.lifecycle.lifecycleScope
+import androidx.media3.exoplayer.source.ClippingMediaSource.IllegalClippingException.Reason
+import androidx.navigation.NavDirections
+import com.aditya.socialguru.BottomNavigationBarDirections
 import com.aditya.socialguru.MainActivity
 import com.aditya.socialguru.R
 import com.aditya.socialguru.data_layer.model.Resource
@@ -23,8 +26,8 @@ import com.aditya.socialguru.domain_layer.helper.Helper.observeFlow
 import com.aditya.socialguru.domain_layer.helper.gone
 import com.aditya.socialguru.domain_layer.helper.myShow
 import com.aditya.socialguru.domain_layer.helper.runOnUiThread
+import com.aditya.socialguru.domain_layer.helper.safeNavigate
 import com.aditya.socialguru.domain_layer.helper.setSafeOnClickListener
-import com.aditya.socialguru.domain_layer.manager.MyLogger
 import com.aditya.socialguru.domain_layer.remote_service.AlertDialogOption
 import com.aditya.socialguru.domain_layer.remote_service.DeleteAccountPasswordResult
 import com.aditya.socialguru.domain_layer.service.SharePref
@@ -41,8 +44,10 @@ class SettingFragment : Fragment(), AlertDialogOption, DeleteAccountPasswordResu
     private val binding get() = _binding!!
 
     private val tagSetting = Constants.LogTag.Setting
+    private var defaultDialogOption: SettingDialogOption = SettingDialogOption.OnResetPassword
 
     private var myLoader: MyLoader? = null
+
     private val settingViewModel by viewModels<SettingViewModel>()
 
 
@@ -121,6 +126,28 @@ class SettingFragment : Fragment(), AlertDialogOption, DeleteAccountPasswordResu
                 }
 
             }.launchIn(this)
+
+            settingViewModel.resetPassword.onEach { response ->
+                when (response) {
+                    is Resource.Success -> {
+                        hideDialog()
+                        response.hasBeenMessagedToUser = true
+                        showSnackBar("A password reset link has been sent to your current email address.", isSuccess = true)
+                    }
+
+                    is Resource.Loading -> {
+                        showDialog()
+                    }
+
+                    is Resource.Error -> {
+                        hideDialog()
+                        response.hasBeenMessagedToUser = true
+                        showSnackBar(response.message.toString())
+                    }
+                }
+
+            }.launchIn(this)
+
         }
     }
 
@@ -145,11 +172,26 @@ class SettingFragment : Fragment(), AlertDialogOption, DeleteAccountPasswordResu
             handleDeleteAccount()
         }
 
+        cardUpdateEmailSetting.setSafeOnClickListener {
+            openBottomSheet(Constants.BottomSheetOpenFor.EmailUpdate)
+        }
+        cardChangePasswordSetting.setSafeOnClickListener {
+            openBottomSheet(Constants.BottomSheetOpenFor.PasswordChange)
+        }
+        cardResetPasswordSetting.setSafeOnClickListener {
+            handleResetPassword()
+        }
+
         switchRemoveStory.setSafeOnClickListener {
 
             // This help to handle or change work manager status
             val intValueOfChecked = if (switchRemoveStory.isChecked) 1 else 0
-            requireActivity().sendBroadcast(Intent(Constants.AppBroadCast.StoryChange.name).putExtra(Constants.DATA,intValueOfChecked))
+            requireActivity().sendBroadcast(
+                Intent(Constants.AppBroadCast.StoryChange.name).putExtra(
+                    Constants.DATA,
+                    intValueOfChecked
+                )
+            )
             updateUserSetting(switchRemoveStory.isChecked)
         }
 
@@ -159,9 +201,24 @@ class SettingFragment : Fragment(), AlertDialogOption, DeleteAccountPasswordResu
 
     }
 
+
     private fun handleDeleteAccount() {
+        defaultDialogOption = SettingDialogOption.OnAccountDelete
         AlertDialog(
             "Are you sure delete your account ?",
+            this@SettingFragment,
+            isForShowDelete = false
+        ).show(
+            childFragmentManager,
+            "MyAlertDialog"
+        )
+
+    }
+
+    private fun handleResetPassword() {
+        defaultDialogOption = SettingDialogOption.OnResetPassword
+        AlertDialog(
+            "A password reset link will be sent to your current email address. Are you sure you want to proceed with resetting your password?",
             this@SettingFragment,
             isForShowDelete = false
         ).show(
@@ -176,7 +233,8 @@ class SettingFragment : Fragment(), AlertDialogOption, DeleteAccountPasswordResu
             pref.getPrefUser().first()?.let {
                 it.userSetting?.let { setting ->
                     binding.apply {
-                        switchRemoveStory.isChecked = setting.isStoryRemoveAfter24HourActive ?: false
+                        switchRemoveStory.isChecked =
+                            setting.isStoryRemoveAfter24HourActive ?: false
                         switchLastSeen.isChecked = setting.isMyOnlineStatusHideEnable ?: false
                     }
                 }
@@ -201,8 +259,10 @@ class SettingFragment : Fragment(), AlertDialogOption, DeleteAccountPasswordResu
         lifecycleScope.launch {
             pref.getPrefUser().first()?.let { user ->
                 val updatedSettings = user.userSetting?.copy(
-                    isStoryRemoveAfter24HourActive = isStoryRemoveAfter24HourActive ?: user.userSetting?.isStoryRemoveAfter24HourActive,
-                    isMyOnlineStatusHideEnable = isMyOnlineStatusHideEnable ?: user.userSetting?.isMyOnlineStatusHideEnable
+                    isStoryRemoveAfter24HourActive = isStoryRemoveAfter24HourActive
+                        ?: user.userSetting?.isStoryRemoveAfter24HourActive,
+                    isMyOnlineStatusHideEnable = isMyOnlineStatusHideEnable
+                        ?: user.userSetting?.isMyOnlineStatusHideEnable
                 ) ?: UserSetting(
                     isStoryRemoveAfter24HourActive = isStoryRemoveAfter24HourActive ?: false,
                     isMyOnlineStatusHideEnable = isMyOnlineStatusHideEnable ?: false
@@ -210,6 +270,12 @@ class SettingFragment : Fragment(), AlertDialogOption, DeleteAccountPasswordResu
                 updatedSettings.let { settingViewModel.updateUserSetting(it) }
             }
         }
+    }
+
+    private fun openBottomSheet(reason: Constants.BottomSheetOpenFor) {
+        val direction: NavDirections =
+            BottomNavigationBarDirections.actionGlobalEmailPasswordChangeDialog(reason.name)
+        navController.safeNavigate(direction, Helper.giveUpAndBottomAnimationNavOption())
     }
 
     private fun showDialog() {
@@ -241,16 +307,29 @@ class SettingFragment : Fragment(), AlertDialogOption, DeleteAccountPasswordResu
     }
 
     override fun onResult(isYes: Boolean) {
-        if (isYes) {
-            DeleteAccountPasswordDialog(
-                this@SettingFragment
-            ).show(
-                childFragmentManager,
-                "MyAlertDialog"
-            )
+        when (defaultDialogOption) {
+            SettingDialogOption.OnAccountDelete -> {
+                if (isYes) {
+                    DeleteAccountPasswordDialog(
+                        this@SettingFragment
+                    ).show(
+                        childFragmentManager,
+                        "MyAlertDialog"
+                    )
+                }
+            }
+
+            SettingDialogOption.OnResetPassword -> {
+               /* showSnackBar(
+                    "A password reset link has been sent to your current email address.",
+                    isSuccess = true
+                )*/
+                if (isYes){
+                    settingViewModel.resetPassword(null)
+                }
+            }
         }
     }
-
 
 
     override fun onResult(isSuccess: Boolean, errorMessage: String?) {
@@ -270,6 +349,11 @@ class SettingFragment : Fragment(), AlertDialogOption, DeleteAccountPasswordResu
         super.onDestroyView()
     }
 
+}
+
+private enum class SettingDialogOption {
+    OnAccountDelete,
+    OnResetPassword
 }
